+++ DuoArtifacts/Artifacts/DizzyPeriArtifact.cs
@@ -1,43 +1,10 @@
-﻿using Shockah.Kokoro;
-using System;
-using System.Collections.Generic;
+﻿using System;
+using Shockah.Kokoro;
 
 namespace Shockah.DuoArtifacts;
 
 public sealed class DizzyPeriArtifact : DuoArtifact, IKokoroApi.IV2.IStatusLogicApi.IHook, IKokoroApi.IV2.IHookPriority
 {
-	public override void OnTurnStart(State state, Combat combat)
-	{
-		base.OnTurnStart(state, combat);
-
-		var actions = new List<CardAction>();
-
-		{
-			var toSubtract = Math.Clamp(state.ship.Get(Status.overdrive), 0, state.ship.Get(Status.shield));
-			if (toSubtract > 0)
-				actions.Add(new AStatus
-				{
-					status = Status.shield,
-					statusAmount = -toSubtract,
-					targetPlayer = true,
-					artifactPulse = Key(),
-				});
-		}
-		{
-			var toSubtract = Math.Clamp(state.ship.Get(Status.perfectShield), 0, state.ship.Get(Status.overdrive));
-			if (toSubtract > 0)
-				actions.Add(new AStatus
-				{
-					status = Status.overdrive,
-					statusAmount = -toSubtract,
-					targetPlayer = true,
-					artifactPulse = Key(),
-				});
-		}
-		
-		combat.QueueImmediate(actions);
-	}
-
 	public double HookPriority
 		=> double.MinValue;
 
@@ -57,7 +24,7 @@ public sealed class DizzyPeriArtifact : DuoArtifact, IKokoroApi.IV2.IStatusLogic
 		args.Combat.QueueImmediate(new AStatus
 		{
 			status = Status.overdrive,
-			statusAmount = overshield,
+			statusAmount = 1,
 			targetPlayer = true,
 			artifactPulse = Key(),
 		});